requirements.txt
# Translation Accuracy Test Tool - Dependencies

# Document Processing
python-pptx>=0.6.21
PyMuPDF>=1.23.0

# Text-to-Speech (TTS)
google-cloud-texttospeech>=2.14.0
edge-tts>=6.1.0

# API Communication
requests>=2.31.0
aiohttp>=3.9.0

# Accuracy Evaluation
jiwer>=3.0.0
sacrebleu>=2.3.0

# Language Processing
langdetect>=1.0.9

# Data Processing
pandas>=2.0.0
